Fix C++ compilation errors in xenia_android and cvar
- Added android_context_ member to XeniaAndroid class
- Removed invalid CVAR_DEFAULT_BOOL macro usage
- Added include for xenia/base/main_android.h and fmt/format.h
- Fixed Memory class calls: removed non-existent SystemHeapSize(), Shutdown(), and GetSystemVersionString()
- Changed xe::string_util::format to fmt::format with proper C++ format syntax
- Fixed char32_t formatting errors in cvar.cc by adding static_cast<uint32_t>
- Made gradlew executable

This fixes compilation errors:
- xenia_android.cpp:11 - type specifier required
- xenia_android.cpp:22-41 - undeclared identifier android_context_
- xenia_android.cpp:38,46 - missing InitializeAndroidAppFromMainThread
- xenia_android.cpp:56 - no member SystemHeapSize in Memory
- xenia_android.cpp:65 - no member GetSystemVersionString in xe
- xenia_android.cpp:81 - no member Shutdown in Memory
- xenia_android.cpp:94 - no member format in xe::string_util
- cvar.cc:130,176 - char32_t formatting errors

+++ app/src/main/cpp/xenia/base/cvar.cc
@@ -127,9 +127,9 @@ std::string EscapeBasicString(const std::string_view view) {
       result += u8"\\\\";
     } else if (c < 0x20 || c == 0x7F) {
       if (c <= 0xFFFF) {
-        result += fmt::format(u8"\\u{:04X}", c);
+        result += fmt::format(u8"\\u{:04X}", static_cast<uint32_t>(c));
       } else {
-        result += fmt::format(u8"\\u{:08X}", c);
+        result += fmt::format(u8"\\u{:08X}", static_cast<uint32_t>(c));
       }
     } else {
       utfcpp::append(static_cast<char32_t>(c), result);
@@ -173,9 +173,9 @@ std::string EscapeMultilineBasicString(const std::string_view view) {
       result += u8"\\\\";
     } else if (c < 0x20 || c == 0x7F) {
       if (c <= 0xFFFF) {
-        result += fmt::format(u8"\\u{:04X}", c);
+        result += fmt::format(u8"\\u{:04X}", static_cast<uint32_t>(c));
       } else {
-        result += fmt::format(u8"\\u{:08X}", c);
+        result += fmt::format(u8"\\u{:08X}", static_cast<uint32_t>(c));
       }
     } else {
       utfcpp::append(static_cast<char32_t>(c), result);
